Move upgrades package to internal package
The upgrades package doesn't need to be accessed via the api and
therefore is subject to change without being versioned.

The code change is mechanical by just moving the package to the
internal package.
