+++ cmd/containeragent/unit/agent.go
@@ -40,8 +40,8 @@ import (
 	"github.com/juju/juju/core/machinelock"
 	"github.com/juju/juju/core/paths"
 	"github.com/juju/juju/internal/upgrade"
+	"github.com/juju/juju/internal/upgrades"
 	jnames "github.com/juju/juju/juju/names"
-	"github.com/juju/juju/upgrades"
 	jujuversion "github.com/juju/juju/version"
 	jworker "github.com/juju/juju/worker"
 	"github.com/juju/juju/worker/introspection"
