+++ cmd/jujud/agent/machine/manifolds.go
@@ -36,8 +36,8 @@ import (
 	internalobjectstore "github.com/juju/juju/internal/objectstore"
 	proxyconfig "github.com/juju/juju/internal/proxy/config"
 	"github.com/juju/juju/internal/servicefactory"
+	"github.com/juju/juju/internal/upgrades"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/upgrades"
 	jworker "github.com/juju/juju/worker"
 	"github.com/juju/juju/worker/agent"
 	"github.com/juju/juju/worker/agentconfigupdater"
