+++ internal/migration/precheck_test.go
@@ -21,9 +21,9 @@ import (
 	environscloudspec "github.com/juju/juju/environs/cloudspec"
 	"github.com/juju/juju/internal/migration"
 	"github.com/juju/juju/internal/tools"
+	"github.com/juju/juju/internal/upgrades/upgradevalidation"
 	"github.com/juju/juju/provider/lxd"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/upgrades/upgradevalidation"
 )
 
 var (
