+++ internal/upgrades/environconfig_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/state"
 	coretesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
+	"github.com/juju/juju/internal/upgrades"
 )
 
 type upgradeModelConfigSuite struct {
