+++ internal/upgrades/package_test.go
@@ -9,7 +9,7 @@ import (
 	gc "gopkg.in/check.v1"
 )
 
-//go:generate go run go.uber.org/mock/mockgen -package mocks -destination mocks/context_mock.go github.com/juju/juju/upgrades Context
+//go:generate go run go.uber.org/mock/mockgen -package mocks -destination mocks/context_mock.go github.com/juju/juju/internal/upgrades Context
 
 func TestPackage(t *stdtesting.T) {
 	gc.TestingT(t)
