+++ internal/upgrades/preupgradesteps_test.go
@@ -12,7 +12,7 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
+	"github.com/juju/juju/internal/upgrades"
 )
 
 type preupgradechecksSuite struct {
