+++ internal/upgrades/upgrade_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/juju/juju/core/model"
 	"github.com/juju/juju/internal/mongo"
 	coretesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
+	"github.com/juju/juju/internal/upgrades"
 	jujuversion "github.com/juju/juju/version"
 )
