+++ worker/upgrader/upgrader_test.go
@@ -32,8 +32,8 @@ import (
 	"github.com/juju/juju/environs/storage"
 	envtesting "github.com/juju/juju/environs/testing"
 	coretools "github.com/juju/juju/internal/tools"
+	"github.com/juju/juju/internal/upgrades"
 	coretesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
 	jujuversion "github.com/juju/juju/version"
 	"github.com/juju/juju/worker/gate"
 	"github.com/juju/juju/worker/upgrader"
