+++ worker/upgradesteps/manifold.go
@@ -14,8 +14,8 @@ import (
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/internal/servicefactory"
+	"github.com/juju/juju/internal/upgrades"
 	"github.com/juju/juju/internal/upgradesteps"
-	"github.com/juju/juju/upgrades"
 	"github.com/juju/juju/worker/gate"
 )
