+++ worker/upgradesteps/package_test.go
@@ -14,9 +14,9 @@ import (
 	gc "gopkg.in/check.v1"
 
 	agent "github.com/juju/juju/agent"
+	"github.com/juju/juju/internal/upgrades"
 	"github.com/juju/juju/internal/upgradesteps"
 	jujutesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
 )
 
 //go:generate go run go.uber.org/mock/mockgen -package upgradesteps -destination clock_mock_test.go github.com/juju/clock Clock
