+++ worker/upgradestepsmachine/manifold.go
@@ -12,7 +12,7 @@ import (
 	"github.com/juju/juju/agent"
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/core/status"
-	"github.com/juju/juju/upgrades"
+	"github.com/juju/juju/internal/upgrades"
 	"github.com/juju/juju/worker/gate"
 )
