+++ worker/upgradestepsmachine/manifold_test.go
@@ -11,8 +11,8 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/agent"
+	"github.com/juju/juju/internal/upgrades"
 	jujutesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
 )
 
 type manifoldSuite struct {
