+++ worker/upgradestepsmachine/package_test.go
@@ -15,9 +15,9 @@ import (
 
 	"github.com/juju/juju/agent"
 	"github.com/juju/juju/core/status"
+	"github.com/juju/juju/internal/upgrades"
 	"github.com/juju/juju/internal/upgradesteps"
 	jujutesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/upgrades"
 )
 
 //go:generate go run go.uber.org/mock/mockgen -package upgradestepsmachine -destination clock_mock_test.go github.com/juju/clock Clock
